Added typechecking, moved the function to cmu_graphics.py as it can be considered core and to be able to translate errors, and fixed test
