+++ tests/image_gen/openImage_util_func.py
@@ -1,3 +1,3 @@
 from cmu_graphics.utils import openImage
-Image(CMUImage(openImage('sample.jpg')), 0,0)
-Image(CMUImage(openImage('https://academy.cs.cmu.edu/static/media/project_10.472f439f.jpg')), 0,0, width=200, height=200)
+Image(CMUImage(openImage('sample.jpg')), 0,0, width=400, height=400)
+Image(CMUImage(openImage('https://academy.cs.cmu.edu/static/media/project_10.472f439f.jpg')), 0,0)
